requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
Django==5.2.3
djangorestframework==3.16.0
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.4
pillow==11.2.1
propcache==0.3.2
pydantic==2.11.5
pydantic_core==2.33.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
yarl==1.20.1
